pyproject.toml
[tool.poetry]
name = "trovi"
version = "0.0.1"
description = "A collection of shared artifacts"
authors = ["Chameleon Cloud <contact@chameleoncloud.org>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.10"
djangorestframework = "^3.12.4"
jsonschema = "^4.2.1"
jsonpatch = "^1.32"
cmarkgfm = "^0.6.0"
keystoneauth1 = "^4.4.0"
python-jose = "^3.3.0"
python-keycloak-client = "^0.2.3"
drf-extensions = "^0.7.1"
mysqlclient = "^2.1.0"
PyJWT = "^2.3.0"
drf-spectacular = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
Faker = "^10.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
